% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gapit_phenotypes_in_folder}
\alias{gapit_phenotypes_in_folder}
\title{Identify phenotype names from GAPIT results in a folder.}
\usage{
gapit_phenotypes_in_folder(path = ".", model = "CMLM")
}
\arguments{
\item{path}{File path to the csv files that GAPIT has created, a character
string.}

\item{model}{Model type used in GAPIT runs, a character string.}
}
\value{
A vector of phenotype names.
}
\description{
Creates a vector of phenotype names from GAPIT results.
}
\examples{
\dontrun{gapit_phenotypes_in_folder()}
\dontrun{gapit_phenotypes_in_folder(path = "usr/gapit_results")}

}
