% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{load_GAPIT_GWAS_all}
\alias{load_GAPIT_GWAS_all}
\title{Load GAPIT Results Tables}
\usage{
load_GAPIT_GWAS_all(path = ".", phenotype, model = "CMLM")
}
\arguments{
\item{path}{File path to the csv files that GAPIT has created, a character
string.}

\item{phenotype}{Phenotype name used in GAPIT, a character string.}

\item{model}{Model type used in GAPIT runs, a character string.}
}
\description{
Load GAPIT Results Tables
}
\note{
To create a vector of phenotype names, use the
\code{\link{gapit_phenotypes_in_folder}} function.
}
