% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{load_GAPIT_GWAS_all}
\alias{load_GAPIT_GWAS_all}
\title{Load GAPIT Results Tables}
\usage{
load_GAPIT_GWAS_all(path = ".", phenotype, model = "CMLM")
}
\arguments{
\item{path}{File path to the csv files that GAPIT has created, a character
string.}

\item{phenotype}{Phenotype name used in GAPIT, a character string.}

\item{model}{Model type used in GAPIT runs, a character string.}
}
\value{
A list containing the five GAPIT csv outputs: PRED, GWAS.Results,
Df.tValue.StdErr, ROC, and Log.
}
\description{
Given a path to a directory and a specific phenotype, this function reads in
GAPIT output csvs and returns a list of these data frames.
}
\note{
To create a vector of all phenotype names, use the
\code{\link{gapit_phenotypes_in_folder}} function.
}
